Added separate project for isolated unit tests
