+++ tests/ArrowDbCore.Tests.Unit.Isolated/xunit.runner.json
@@ -1,0 +1,7 @@
+{
+    "$schema": "https://xunit.net/schema/current/xunit.runner.schema.json",
+    "diagnosticMessages": true,
+    "parallelizeAssembly": false,
+    "parallelizeTestCollections": false,
+    "showLiveOutput": true
+}
