+++ tests/ArrowDbCore.Tests.Unit/TrackingVariables.cs
@@ -2,23 +2,6 @@
 
 public class TrackingVariables {
     [Fact]
-    [Trait("Category", "Isolated")]
-    public void Instance_Counter_Is_Zero_At_Startup() {
-        Assert.Equal(0, ArrowDb.RunningInstances);
-    }
-
-    [Fact]
-    [Trait("Category", "Isolated")]
-    public async Task Instance_Ids_Match_Running() {
-        const int count = 10;
-        var dbs = new ArrowDb[count];
-        for (var i = 0; i < count; i++) {
-            dbs[i] = await ArrowDb.CreateInMemory();
-        }
-        Assert.Equal(count, ArrowDb.RunningInstances);
-    }
-
-    [Fact]
     public async Task Changes_Match_Additions() {
         const int count = 10;
         var db = await ArrowDb.CreateInMemory();
